Move CheckWidths to its own file
